changes done

+++ controller/user.controller.js
@@ -1,10 +1,15 @@
+const mongoose = require('mongoose')
 const User = require('../models/User')
 
 // get all users 
 exports.user_get = async (req,res)=>{
     try{
         const user = await User.find();
-        res.send(user)
+        if(user){
+            res.json({data:user,status:200})
+        }else{
+            res.json({message:"Users not exist!",status:404})
+        }
     }catch(err){
         res.status(500).json(err);
     }
@@ -13,8 +18,17 @@ exports.user_get = async (req,res)=>{
 // get user by id 
 exports.user_get_by_id = async (req,res)=>{
     try{
-        const user = await User.findById(req.params.id);
-        res.send(user)
+        var isValid = mongoose.Types.ObjectId.isValid(req.params.id)
+        if(isValid){
+            const user = await User.findById(req.params.id);
+            if(user){
+                res.json({data:user,status:200})
+            }else{
+                res.json({message:"User Id not exist!",status:404})
+            }
+        }else{
+            res.json({message:"User Id not valid!",status:400})
+        }
     }catch(err){
         res.status(500).json(err);
     }
@@ -25,12 +39,11 @@ exports.user_post = async (req,res)=>{
     try{
         const exit_user = await User.find({email:req.body.email});
         if(exit_user.length>0){
-            res.json({message:"User Email already existed!"})
-            return true;
+            res.json({message:"User already exist!",status:400})
         }else{
             const new_user = new User(req.body)
             const user = await new_user.save();
-            res.json(user)
+            res.json({data:user,status:200})
         }
     }catch (err) {
         res.status(500).json(err);
@@ -40,14 +53,34 @@ exports.user_post = async (req,res)=>{
 // update user by id 
 exports.user_update_by_id = async (req,res)=>{
     try{
-        const exit_user = await User.findById(req.params.id);
-        if(!exit_user){
-            res.json({message:"User id not existed!"})
-            return true;
+        var isValid = mongoose.Types.ObjectId.isValid(req.params.id)
+        if(isValid){
+            const exit_user = await User.findById(req.params.id);
+            if(!exit_user){
+                res.json({message:"User id not exist!",status:404})
+            }else{
+                const user = await User.findByIdAndUpdate(req.params.id,req.body)
+                let suser = await user.save();
+                res.json({data:suser,status:200})
+            }
         }else{
-            const user = await User.findByIdAndUpdate(req.params.id,req.body)
-            await user.save();
-            res.json(user)
+            res.json({message:"User Id not valid!",status:400})
+        }
+    }catch (err) {
+        res.status(500).json(err);
+    }
+}
+
+// update user by email
+exports.user_update_by_params = async (req,res)=>{
+    try{
+        const exit_user = await User.find({email:req.body.email});
+        if(exit_user.length<1){
+            res.json({message:"User id not existed!",status:404})
+        }else{
+            const user = await User.findOneAndUpdate(req.body.email,req.body)
+            const suser = await user.save();
+            res.json({data:suser,status:200})
         }
     }catch (err) {
         res.status(500).json(err);
@@ -57,14 +90,11 @@ exports.user_update_by_id = async (req,res)=>{
 // delete all user
 exports.user_delete = async (req,res)=>{
     try{
-        const exit_user = await User.dele(req.params.id);
-        console.log(exit_user)
-        if(!exit_user){
-            res.json({message:"User id not existed!"})
-            return true;
+        if(req.body.all=="allUsers"){
+            const user = await User.deleteMany()
+            res.json({data:user,status:200})
         }else{
-            const user = await User.findByIdAndDelete(req.params.id)
-            res.json(user)
+            res.json({message:"Missing params!",status:400})
         }
     }catch(err){
         res.status(500).json(err);
@@ -74,14 +104,17 @@ exports.user_delete = async (req,res)=>{
 // delete user by id 
 exports.user_delete_by_id = async (req,res)=>{
     try{
-        const exit_user = await User.findById(req.params.id);
-        console.log(exit_user)
-        if(!exit_user){
-            res.json({message:"User id not existed!"})
-            return true;
+        var isValid = mongoose.Types.ObjectId.isValid(req.params.id)
+        if(isValid){
+            const exit_user = await User.findById(req.params.id);
+            if(!exit_user){
+                res.json({message:"User id not exist!",status:404})
+            }else{
+                const user = await User.findByIdAndDelete(req.params.id)
+                res.json({data:user,status:200})
+            }
         }else{
-            const user = await User.findByIdAndDelete(req.params.id)
-            res.json(user)
+            res.json({message:"User Id not valid!",status:400})
         }
     }catch(err){
         res.status(500).json(err);
